Revised main dialog panel layout Show arrow icon for shortcut files and symlinks Execute the "on completion" command asynchronously Resolved invalid grid background when context menu is shown Set negative file time tolerance to disable file time check Optimized sequence of steps when saving database files Prevent temporary incorrect statistics after unexpected increase in workload Fixed default height when mixing panels with and without caption on main dialog New view filter button "show excluded items" New keyboard shortcuts for file filter and sync settings Removed libpng15.so dependency for openSUSE 13.1 Updated help file Updated translation files

+++ FreeFileSync/Source/comparison.h
@@ -45,7 +45,7 @@ struct FolderPairCfg
 std::vector<FolderPairCfg> extractCompareCfg(const MainConfiguration& mainCfg); //fill FolderPairCfg and resolve folder pairs
 
 //FFS core routine:
-void compare(size_t fileTimeTolerance, //max allowed file time deviation
+void compare(int fileTimeTolerance, //max allowed file time deviation
              xmlAccess::OptionalDialogs& warnings,
              bool allowUserInteraction,
              bool runWithBackgroundPriority,
